Update README.md
In this study, we proposed the use of the HODMD for
power quality analysis in electric grid scenario. In the proposed
methodology, the amplitude and frequency for specific features
in the data are extracted by utilizing the computation of
HODMD modes through the low-rank approximation. The
proposed analysis indicated that HODMD is quite efficient in
the detection and description of voltage waveform distortions,
specifically harmonics and inter-harmonics. The technique
successfully pinpointed the prominent frequencies and ampli￾tudes of these distortions, with the resulting HODMD modes
presenting distinct graphical depictions of the spatio-temporal
patterns linked to each distortion. In contrast to alternative
techniques, HODMD boasts numerous benefits including its
ability to handle high-dimensional data, robustness against
disturbances and outliers, and capability to detect nonlinear
dynamics.. The approach is highly efficient from a computational standpoint and can be readily implemented on extensive
data sets. This study emphasizes the capability of HODMD as
a means of analyzing power quality and indicates its potential
as a beneficial resource for engineers and scholars working
in this area. Future directions may investigate how HODMD
can be utilized for analyzing various forms of power system
information, and also delve into creating automated methods
for extracting and categorizing features based on DMD modes