Update README.md
The stability, effectiveness, and high-quality operation of the electrical system are maintained through the monitoring of fundamental parameters like frequency and amplitude. The current research provides a data-driven hybrid technique for the extraction of fundamental parameters using the Higher-order dynamic mode decomposition (HODMD) algorithm in smart grid. HODMD's capacity to account for non-linearities enables it to produce accurate representations of the underlying dynamics, resulting in better model estimation. Thus, HODMD is an appropriate method for power quality analysis due to its capacity to analyze multidimensional data and non-uniformly sampled data. In this work, the potential of HODMD is examined for the estimation of fundamental frequency, amplitude, and the existence of disturbance components, such as harmonics. In the proposed methodology, HODMD modes are computed using the low-rank approximation and then extracted the amplitude and frequency information from the data. The hyper parameters of the proposed methodology are properly tuned to get an accurate estimate of the underlying parameters. The optimistic results on various synthetic and real-time scenarios indicate that the proposed methodology can be utilized to determine the power system frequency and amplitude in general periodic and quasi-periodic dynamics